Finished Floyd Warshall implementation
Added slowestPerson method, this will determine who will take the longest time to traverse any path, this is used to determine the worst case show running time.

Added Implementation for the timeRequiredForCompetition method, it now builds the shortest paths from every vertex to every other vertex and stores them in a 2D array. This array is then traversed to check if any vertex is unreachable. If a vertex is unreachable, return -1, otherwise keep track of what the longest path is. 

Finally convert the distances from kilometers to meters and determine the number of minutes the show will run for. The int value returned is the ceiling of the time required.